try osx sudo workaround

+++ build/before_install.sh
@@ -11,8 +11,10 @@ if [ "$1" == "linux" ]; then
     rm -f /usr/lib/x86_64-linux-gnu/libcrypto.so
     rm -f /usr/lib/x86_64-linux-gnu/libssl.so
 else
-    brew upgrade openssl
-    rm -f /usr/local/opt/openssl/lib/libssl.1.0.0.dylib
-    rm -f /usr/local/opt/openssl/lib/libcrypto.1.0.0.dylib
-
+    if [ "$(id -u)" != "0" ]; then
+        brew update
+        brew install openssl
+        rm -f /usr/local/opt/openssl/lib/libssl.1.0.0.dylib
+        rm -f /usr/local/opt/openssl/lib/libcrypto.1.0.0.dylib
+    fi
 fi
